feat(new-client): Refactored index and platform main files
