+++ src/new-client/src/App/LiveRates/Tile/TearOut/handleTearOut.web.ts
@@ -1,3 +1,4 @@
+import { ROUTES_CONFIG } from "@/constants"
 import { constructUrl } from "@/utils/url"
 import { openWindow } from "@/utils/window/openWindow"
 import { tearOut } from "./state"
@@ -5,7 +6,7 @@ import { tearOut } from "./state"
 export function handleTearOut(symbol: string) {
   openWindow(
     {
-      url: constructUrl(`/spot/${symbol}`),
+      url: constructUrl(ROUTES_CONFIG.tile.replace(":symbol", symbol)),
       name: symbol,
       width: 380,
       height: 170,
