--- src/new-client/src/Finsemble/FinsembleApp.tsx
@@ -1,57 +1,0 @@
-import { BrowserRouter, Route, Switch } from "react-router-dom"
-import { Analytics } from "@/App/Analytics"
-import { LiveRates } from "@/App/LiveRates"
-import { TileView } from "@/App/LiveRates/selectedView"
-import { Trades } from "@/App/Trades"
-import { DocTitle } from "@/components/DocTitle"
-import { BASE_PATH } from "@/constants"
-import { TornOutTile } from "@/App/LiveRates/Tile/TearOut/TornOutTile"
-
-export const FinsembleApp: React.FC = () => (
-  <BrowserRouter basename={BASE_PATH}>
-    <Switch>
-      <Route
-        path="/analytics"
-        render={() => (
-          <DocTitle title="Analytics">
-            <Analytics hideIfMatches={null} />
-          </DocTitle>
-        )}
-      />
-      <Route
-        path="/blotter"
-        render={() => (
-          <DocTitle title="Trades">
-            <Trades />
-          </DocTitle>
-        )}
-      />
-      <Route
-        path="/tiles"
-        render={() => (
-          <DocTitle title="Live Rates">
-            <LiveRates />
-          </DocTitle>
-        )}
-      />
-      <Route
-        path="/spot/:symbol"
-        render={({
-          location: { search },
-          match: {
-            params: { symbol },
-          },
-        }) => {
-          const query = new URLSearchParams(search)
-          const view = query.has("tileView")
-            ? (query.get("tileView") as TileView)
-            : TileView.Analytics
-
-          return (
-            <TornOutTile symbol={symbol} view={view} supportsTearOut={false} />
-          )
-        }}
-      />
-    </Switch>
-  </BrowserRouter>
-)
