+++ src/new-client/src/main.finsemble.ts
@@ -1,1 +1,5 @@
-export { default } from "@/Finsemble"
+import { FinsembleApp } from "@/Finsemble"
+
+export const gaDimension = 'finsemble';
+
+export const getMainApp: () => React.FC = () => FinsembleApp
