+++ src/new-client/src/main.launcher.ts
@@ -1,1 +1,5 @@
-export { default } from "./Launcher"
+import { LauncherApp } from "@/Launcher"
+
+export const gaDimension = 'launcher';
+
+export const getMainApp: () => React.FC = () => LauncherApp
