+++ src/new-client/src/main.openfin.ts
@@ -1,1 +1,5 @@
-export { default } from "@/OpenFin"
+import { OpenFinApp } from "@/OpenFin"
+
+export const gaDimension = 'openfin';
+
+export const getMainApp: () => React.FC = () => OpenFinApp
